Generate absolute URLs to all social media links
Links to sites with no matching icon were being returned as is.
This commit takes proper account of relative paths, so that
'dev.to/user' doesn't end up pointing to 'localhost:8080/dev.to/user',
for example.

Also fixed: the blog post generator was throwing a KeyNotFoundException
in watch mode whenever a file changed

TODO
----
- open a PR adding `Seq.tryFind` to the site template in the Fornax repo